requirements.txt
pandas==2.0.2
numpy
psycopg2-binary
flask==2.0.2
sqlalchemy